pyproject.toml
[project]
name = "mcp-recipe-server"
version = "2.0.0"
description = "MCP Server with Recipe Generation and Food Analysis"
authors = [
    {name = "Akash", email = "smu.akash9@gmail.com"},
]
dependencies = [
    "fastmcp>=1.0.0",
    "openai", 
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "Pillow>=9.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0"
]

[project.scripts]
mcp-recipe-server = "mcp_recipe_server.main:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"